Fixed selected state of virtualenvs action with "Filter" icon
The action "Show virtual environments associated with other projects"
was active, but it was represented by the pressed "Filter" icon, that
usually means a filtered set of the initial results.

This fix changes the name of the action from "Show virtualenvs" to
"Hide virtualenvs" and changes its state to off by default.