[vcs-log] reset log state when initialization is cancelled or fails
If a git executable is improperly configured, log initialization does not work. After correct executable is configured, it needs to be done again. The initialization can be done when one of the tabs is opened, when one of roots is refreshed or when user presses Refresh button.

IDEA-186463

(cherry picked from commit ae5ff6c)