JPS: add extension to provide additional constant affection resolvers
JVM languages such as Kotlin can inline constant values at use-site.
When a constant's value is changed, all use-sites must be recompiled.
Since values are inlined, there are no references in bytecode
so Java builder does not know when a Java constant is referenced in Kotlin.
Kotlin builder knows where each FQ-name is used, but does not know
when a Java constant is changed.

To fix KT-16091, Kotlin plugin needs to provide additional ConstantAffectionResolver,
so it would be notified when a constant value is changed.

In this commit ConstantSearchProvider class is added.
Additional instances of ConstantAffectionResolver provided by ConstantSearchProvider
implementations are used only when "main" ConstantAffectionResolver
(that searches using Intellij) is used (otherwise Java builder forces rebuild,
so it does not make sense).