PY-24728 Follow symlinks inside virtualenv to detect stdlib modules
This way we deal both with symlinks in a venv pointing to actual modules
of its base interpreter and also with rare interpreter layout of
on some x86_64 systems, like Fedora, where there is "lib64"
alias to "lib" directory of a virtualenv, and stdlib of the base
interpreter is also under /usr/lib64 instead of /usr/lib.

On the whole, we can follow this path now:
venv/prefix/lib64/os.py ->
venv/prefix/lib/os.py ->
base/interpreter/prefix/lib64/os.py

and detection of the lib directory of a base interpreter is
always correct since it doesn't depend on the actual name in use.

(cherry picked from commit cb72c40)