[vcs-log] fix spaces between labels in non-compact mode
Grey area width around a label was accidentally dependent on LEFT_PADDING, while it should be a percentage of middle padding, which is the distance between an end of one label text and beginning of next label icon. So when LEFT_PADDING was changed in 053f8ec87d9a3fa1b6e500e52320e15a3dbd8ec9 to accommodate smaller icons, labels in non-compact mode started to be painted incorrectly.