[file-history] use old history algorithm when log is not ready
When log is not ready it provides a data pack that is not full. So we can not filter it to show a full file history. In order to fix that, old file history algorithm is used and a fake data pack is created from it.

IDEA-176197

(cherry picked from commit 76da69e)