fix SIOOBE (EA-117319)
Contrary to other renderers the `PropertyRenderer` builds template for
the whole class, because it needs to insert both getter and setter, and
their common parent is a class. Problem was appearing in case of class
without the closing brace, its marker was not stretched to the right
after insertion of a methods, because methods are inserted to the end of
the class and the marker was not greedy to the right.

(cherry picked from commit 00fa1e2) (IDEA-CR-29987)