Add versionDetailsProvider extension point to the platform.
This extension point allows the extensions to supply
application-specific version details in addition to the platform,
system and JDK information always populated by the platform by
default.

For example, Android Studio would provide additional details like
Gradle and Android Gradle Plugin versions, as well as versions
of some Android SDK components.

This is going to make it easier for users to collect version
information when they file bugs, which apart from the general
UX improvement would also reduce turnaround time as it'll be
less likely that a necessary detail might need to be requested
from the user additionally.

This CL also makes the changes to SendFeedbackAction and
AboutPopup, which are the two sources of the version information
used by people when filing bugs.