PY-28715 Insert "object" as a placeholder for variables with unknown type
This way we don't add an import of typing.Any in advance for a type hint
that is likely to be changed right away, so that user had to remove it
manually afterwards or call optimize imports every time.
If Any is used as a component in some compound type like "List[Any]" we
still add the corresponding import as before.

(cherry picked from commit fb05cee)