lst: track changelist for not-yet-initialized trackers
Fix issue, when file from non-active changelist would briefly jump
to active changelist and back when it's opened in Editor.

PartialTracker is immediately registered in CLM and is used to
determine file's changelist. But until its base content is loaded,
it has no changed blocks and "affected changelists" defaults to active changelist.