build scripts: show block names in exception message
When some build step fails it's not evident to find the real cause from the error message (it may be something like 'exec returned: 2') and because TeamCity doesn't show log for parallel tasks properly (TW-8249) it's also hard to find log messages printed before the exception was thrown. Now the error message will include descriptions of the parent blocks.