PY-28808 Render empty tuple types simply as "Tuple" without brackets
so as not to produce syntactically illegal type hints. It's not clear,
though, whether we should render such types as "tuple" (not capitalized)
in other contexts, say, in Quick Documentation, as we do for built-in
collections parametrized with Any at the moment.

(cherry picked from commit 15c97cf)

(cherry picked from commit abb66a7)