Make sure we enable analytics if allowed
Upstream is migrating the storage for user consent to a new location
(see commit 98b4849fbfec8bb528e3dc292fe1c78958cb4f0b, part of a larger
refactoring to simplify usage tracking and analytics). Instead of
usage.statistics.xml, .local/share/Google/consentOptions/accepted is
now the source of truth, toggled via Help -> Data Sharing Options.

As a result, if we don't import settings from a previous version in
which usage.statistics.xml was present, we risk not initializing the
crash reporter. While we're in this migration phase, we should try to
configure analytics also when querying the consentOptions.

Test: manual
Bug: 72961188
Change-Id: Ia141d5865ede010786d50e29f3a7798ee4f74adb