[groovy] fix IDEA-176850
Before the change all reference expressions inside closures were
resolved against that closure delegation strategy. In case of original
issue that strategy was DELEGATE_ONLY, so reference resolution stopped
after processing the delegate and it never got to process imports,
inners classes and so on.

After the change reference expressions that are not resolved to a
variable are then tried to resolve to a class by its simple name,
reusing logic from code reference resolution.
The same happens in groovy own ResolveVisitor.