PY-28848: Complete python modules in PySymbolFieldWithBrowseButton
If we have no qualified name yet we still provide list
of all possible modules (files and packages with __init__.py) found in module or project and sdk roots.

There is also refactoring that adds extension methods both to VirtualFile and PsiFileSystemItem to check if one is python module. Former does not use latter because of performance.

This functionality is enabled for flask, tests, and plain python script

(cherry picked from commit 4dd92e6)