Add an onStart method to ExternalProjectRefreshCallback
When ExternalSystemUtil#refreshProject is called, a rerun import action is
created and attached to the build view. This action is called when the
"Refresh" button is pressed and what it does is simply call refreshProject
again with the same parameters. For Android Studio, we do some setup
before calling refreshProject, but since this rerun action calls
refreshProject directly, we cannot do this inicial setup.

This CL adds an onStart method to ExternalProjectRefreshCallback which
allows any setup steps to be done before starting project refresh,
causing those steps to be repeated when the refresh button is pressed.