EA-116787 Use smart pointer to find and delete original type comment
Because inserting malformed type hint as annotation can alter AST so
drastically that the original comment becomes invalid. The result of
the intention is still syntactically illegal but at least we don't throw
exceptions anymore.

Attempting to parse every single type comment to check its validity
seems somewhat wasteful. We'd better come up with a decent heuristic for
that purpose.

The original case for a type comment combining type hint with subsequent
plain text part has been already handled thanks to changes for
PY-21195, as we stopped inserting these trailing comments in annotations
themselves, but I added a test for it just in case.