PY-28199 Improve transformation of camel case into snake case
Namely, an underscore is inserted in front of an uppercase letter
only if it's either preceded by a lowercase letter or digit or it's
the last letter in a block of uppercase letters followed by a lowercase
one, so that, say, "UUIDKey" is turned into "uuid_key".
Existing series of underscores are not extended.

The same convention has been used in a similar quickfix for Ruby.