[jvm-lang] add JvmPrimitiveTypeKind enum
This eliminates dependency on PsiType.XXX types in JVM code, and
provides ability to use custom JvmPrimitiveType-s as long as their kind
is correct.

Some plugins use PsiPrimitiveType with arbitrary name to represent fake
types (since it was convenient not to implement all PsiType methods).
This ability was preserved but deprecated.