Refactor studio patcher to depend on javaxdelta jar
The main reason behind this refactor is we need to check in the code into
google3's third_party repo. Due to the one version policy, we can't check
in the code in its current state because it contains source from javaxdelta.

I updated the code to depend on javaxdelta-2.0.1.jar. fastutil-7.2.1 is a
runtime dependency of javaxdelta.

BUG: b/76162001
Test: n/a
Change-Id: I3110d82c841731dc60ffee3a1c4ab6036c78d229