PY-29174: Install framework in modal window for local and remote sdk
We can check is framework is installed for local sdk, but not for
remote.

So, fo local we call ``installFrameworkIfNeeded`` with
forceInstall and for remote is checks if framework is installed
by itself.

``installFrameworkIfNeeded`` shows modal window while
checking/installing framework (both local and remote) and then
runs callback to generate project and do further actions

Done for Flask and Django.

This is not supported for plugin for now