IDEA-181823 Don't make moved file unversioned
Initially there was IDEA-153272 (move a file overwriting an existing file => both files are deleted) which was fixed in 783963b.

That fix introduced IDEA-118140 (move untracked file => file is staged), which was fixed in a30e7bb.

However, that fix introduced IDEA-181823 (move versioned file => sometimes it stays unversioned) for the following reason:

When a new file is created (including moves), in addition to the VcsVfsListener it is also handled by the asynchronous GitUntrackedFilesHolder. It sees the new file and adds it as possibly unversioned. Then the status is requested for this file, Git says that is is unversioned indeed (because no git add was executed yet), and the FileStatus becomes UNKNOWN. Then, at the end of the Command, the control flow gets to the VcsVfsListener, which doesn't attempt to add the file, since its status has already been defined as unversioned.

Current fix basically reverts a30e7bb and 783963b, and fixes the original issue with overwrite (IDEA-153272) in a following different way:

If a file is moved on top of another file (overwrite), the VFS at first deletes the original file, and this deletion causes `git rm` to be called. But in fact the VCS doesn't need to handle this deletion at all: it is just a trick to implement the overwrite. Therefore, the fix is to remove the unversioned moved file from the list of deleted files. Versioned moves are already handled in doNotDeleteAddedCopiedOrMovedFiles() at the end of the Command, but unversioned moves can't be handled there because they are not recorded.

(cherry picked from commit 7760397)