IDEA-189350 lst: disable UndoableAction registering for Redo actions
These actions can cause different hard-to-debug issues in UndoManager.
One of the known issues is mishandling of transparent actions.

Redo action is being registered only at the end of command.
If the document was modified in a transparent command, redo action will not
be registered until when next (completely unrelated) command is finished.

As a result, it causes "Can't Undo" errors, if that other command was
not supposed to be undoable (and was performed on workspace.xml or
some lightweight temporal file)

(cherry picked from commit 4a34389)
IDEA-CR-31263