[file-history] use correct key for to check if history is in progress
In order to avoid computing history twice, a special handler is used, where each history process registers itself with a special key. Previously, a single key was used for all files, so we could not compute histories for two different files simultaneously. In order to fix that, a key specific to each file is used.

IDEA-148366