IDEA-189376 lst: do not move files into default changelist on update
We used to put all changes into mappings in worker from ChangeListUpdater,
even if there are a partial tracker registered for this path.
As such changes are not included into myChangesBeforeUpdateMap mapping,
they are always mapped into default list.
If the tracker was released during the CLM update, we would use changelist
from updatedWorker - and move file into default changelist (instead of
the one, that was assigned to this file in `unregisterChangeTracker`).

There is still an issue, if the tracker was registered, modified
and destroyed during the update (so updatedWorker knows nothing about it).