Fix ClassCastException in ExceptionRegistry
Issue:
As exceptions are added to ExceptionRegistry, the tree of StackFrames
(inner nodes) and LeafFrames (leaf nodes) is formed. LeafFrame is a
subclass of StackFrame with additional data (count and timestamp).

When an exception is added which is the same as an exception
already added to the registry, but with some top frames removed,
the code casts an inner node to a leaf node. This results in a
ClassCastException.

Fix:
LeafFrame (now renamed to LeafNode) no longer inherits from a
StackFrame. Instead, LeafNode is now a property of a StackFrame.

Code cleanup:
 * This change simplifies the implementation of some methods
   (addChild, addLeaf).
 * Removed count property from inner frames (it was never used)
 * LeafFrame renamed to LeafNode as it no longer stores frame data.

Change-Id: I1d41016e451d67c858141d20c1a6ea6fd6341a8b
Fixes: 77499672
Test: Added ExceptionRegistryTest.testInsertLeafIntoExistingChain