Revert "Revert Change I421cb797"
This reverts commit 405282aac8dc2a6d0c85bfbc1a4da470e1c72d29.

I421cb797 re-introduces the ASWB builds that were disabled for the
2018.1 merge. It was reverted because it caused all UI tests to start
failing. This CL re-introduces the builds again, except it does not
make the uitest target depend on the ASWB builds.

The ASWB dependency will be added to the uitest target in a separate
topic, along with other changes that get our UI tests working again.

Bug: 76461262
Test-Strategy: all
Test: UI Tests
Change-Id: I7b62c203b8b20ee8823ed518df515d546d033e82