Allow specifying ApplicationStarter via system property
The IDE selects the ApplicationStarter whose getCommandName() matches
the first command-line argument. To run UI tests starting the IDE with a
native launcher, an alternate mechanism must be provided, as there's no
way to specify command-line arguments.

This CL introduces the idea.application.starter.command system property.
The command line arguments will be checked first, falling back to the
system property, which could be set in a VM options file.

Bug: 77820421
Change-Id: I5b50cefa76e298855bd0836c9a97a7b2cbbaa430