Fix UI test bundling after great module rename
In the great module rename, the map between production-scope JAR names
and module names was transformed by a module renaming function, while
the corresponding map for test-scope modules was not. Keys of each map
are used to index the other, which led to breakage.

This CL applies the same transformation to the test-scope modules map.

Test: ran build_studio.sh --uitests on linux workstation
Change-Id: Iefd3f6789d8d49ad3791ddcffd099a1a63b4cdac