Set Android Studio defaults in UpdateOptions
Android Studio prefers to turn on update check as the default, and to
use EAP as the default channel for canary builds.  The field is
changed to BETA or RELEASE when building beta, RC, stable in release
branch.

Note: there is AndroidStudioUpdateStrategyCustomization that sets
forceEapUpdateChannelForEapBuilds to false, because making it true
appears to cause further tweaking than just making EAP the default
for EAP builds.

Also, Update RELEASE.md step about correctly setting the default
update channel for canary, beta, RC, stable builds.

Also, Fix a couple of formatting issues in markdown of RELEASE.md.

Bug: 77861318
Test: manual
Change-Id: I27114e503e9192e97681327f47e602cc9676d986