CPP-5797: mixdebug: Stopping Python using breakpoint on function-trap
Calling the special interpreter function type_set_bases after setting
the breakpoint during python pause in order to gdb stop gdb. The signal
had been used earlier but not sure if it was safe way. Also this trap
method stops the thread, that reached the breakpoint on python unlike
stopping with signal.