Make uicoverage target depend on build-aswb
NlEditorTest#basicLayoutEdit, the only test in the SANITY group which
targets a Bazel project, is failing because the Bazel plugin isn't
available. This is because the uicoverage target in build.xml doesn't
depend on build-aswb, the target responsible for building and
including the Bazel plugin.

This CL resolves the issue by adding the missing dependency.

Bug: 77696373
Test: Will monitor post-submit results
Change-Id: I829fecb2f0bac83b0207c99f79ede8d9097fe1de