Resize myOriginalDocument to a message boundary
The original implementation can leave a partial log message at the head of
the myOriginalDocument buffer when it resizes it. The partial message breaks
the Android Studio message parsers when the Logcat console is refiltered.
Resize the buffer to a message boundary (the terminating newline) so the
parsers don't break.

Bug: https://issuetracker.google.com/77876440