[vcs-log] fix recreating log after reloading mappings from disk
When vcs mappings are changed on disk, ProjectLevelVcsManager.readDirectoryMappings first clears the old mappings completely. This triggers VCS_CONFIGURATION_CHANGED event. Then new mappings are set, which triggers the event again. So recreateLog is called twice, the first time it just disposes the old log, the second it should create a new one.

The issue was that the callback creating a new log was only called if an old log was present (after disposing it). Which led to new log not being created. Fix is to always invoke the callback.

(cherry picked from commit bfa074b)