PY-29471 Properly delete temp directory created by means of tempdir module
Previously in case of error pip run with runpy.run_module() terminated
the execution of the script abruptly with SystemExit and this cleanup
wasn't performed.

Normally it didn't happen though, since temp directory was created on
Java side in PyPackageManagerImpl.install() that was responsible for
its removal.

(cherry picked from commit d280c27)