Use a warning instead of error icon on build view for warnings
We currently generate a FailureResult when there are notifications
during sync, without checking if those notifications are warnings or
errors. That causes the icon to be the same for failed syncs and
successful syncs with warnings. This was done in purpose since the build
tool window is not currently processing the warnings from SuccessResult,
and with it, none of the quickfixes and hyperlinks that come with the
notifications.

This CL checks if there are only warnings on a FailureResult to decide
what icon to show.

Bug: 77278972
Test: Manually caused warning and failures to confirm the correct icon
is shown and that the notification are also present.

Change-Id: I519f12c4046d6505d83004e1427eb5b1d49bd3fb