[git] preserve order of merge commits when necessary
For merge commits that are not different to one of the parents collecting changes requires additional "git log" call in order to get tree hashes. In order to avoid multiple extra git invocations, these calls were bundled together. This led to merge commits appear at the end of the list, which was fine for indexing purposes, but led to problems in other cases, such as branches comparison. This commit introduces a parameter that allows to preserve commits order.

Issue was introduced in eed01c212391c57bad9e5c3eba610f5916990d39.

IDEA-186772