IDEA-191309 git: update staging area using "--cached" argument on commit
fix "error: the following file has local modifications" on attempt to
call `git rm -- <before-rename path>` on a renamed file, excluded from commit.
This might be caused by git pre-commit-hook, that created file with
that path during the commit.
(so no now it is a modification + new file, and not a rename).