IDEA-191309 git: update staging area using "git add -A" on commit
Unify logic for deleted, new, modified, renamed, etc files.
This prevents possible "Deleted-Unversioned" file status when
we expected file to be removed, but it is actually a modification on disk.
(ex: because pre-commit hook recreated a file that was deleted before)

Use "-A" as a parameter for "addPaths", to unify logic between git versions.
(This flag is set by default since git 2.0.0)