svn: Optimize "SvnVcs.filterUniqueRoots()" (IDEA-140811)
Method could be called under write action (for instance, when new modules are added to the project). And it is called lots of times during project import process. So svn client calls are replaced instead with searching for working copies in VFS.

Previous logic is still available if "svn.filter.unique.roots.by.url" registry key is "true".