PY-29779 Update virtualenv to 15.2.0
On Python < 3.3 we rely on virtualenv script to update setuptools and
pip in created virtualenvs instead of using those we bundle. On the
affected version of MacOS virtualenv failed to download newer versions
of the packages and, thus, fell back to pip 9.0.1 (not able to handle
newer TLS on pypi.org) that it's shipped with itself.

(cherry picked from commit 48ade29)

(cherry picked from commit 86f5b37)

(cherry picked from commit 506e588)