don't invalidate all PSI on root change
It was originally done to avoid UI freeze (IDEA-172762),
but has led to too many PIEAEs (like IDEA-191185, IDEA-188292, IDEA-184186, EA-114990).

Ideally those clients should all be converted to smart pointers, but that proved to be quite hard to do, especially without breaking API.
And they mostly worked before those batch invalidations.

So now we have a smarter way of dealing with this issue. On root change, we mark
PSI as "potentially invalid", and then, when someone calls "isValid"
(hopefully not for all PSI and hopefully in a background thread),
we check if the old PSI is equivalent to the one that would be re-created in its place. If yes, we return valid. If no, we invalidate the old PSI forever and return the new one.