svn: Do not check parent directories when auto detecting svn
We do not look for administrative directories in parents for other vcses (like git, hg). And so prior to this commit we could get confusing results for instance when directory is in git repository, but there is some ".svn" subdirectory (even not real working copy) somewhere in the parent hierarchy.

IDEA-115100