Record event timings
Instrumented IdeEventQueue with a timer that records the time taken
to dispatch individual events. This data is collected in a histogram
and sent back periodically in the StudioPerformanceStats message.

The histogram data is collected using the HdrHistogram library, which
has been added as a new gradle dependency.

Nested event loops often occur when opening modal dialogs - that is,
one or more other events may be dispatched recursively while still
processing another event. When that occurs, the time spent waiting for
and dispatching the nested events does not count toward the time spent
on the outer event - otherwise, a long UI freeze would be recorded
every time a modal dialog box is opened.

To handle this sort of recursive timing, we've introduced the notion
of a RecursiveStopwatch, which is basically a stopwatch that allows
its state to be saved and restored. When processing a nested event loop,
we save the stopwatch state immediately before timing the nested event
and restore the stopwatch state immediately afterward.

Bug: 78896974
Test: Unit tests included
Change-Id: I181d373b6e3aedab25c8b65b1c880b1321e38643