PY-26492 Show the base interpreter combo box instead of Python version field
The user might want to specify a custom path to an interpreter the same
way they do with `pipenv --python /path/to/python`. We removed the
ability to specify just a version of Python since Pipenv finds a
corresponding interpreter on `$PATH` anyway and we do exactly that in
our base interpreter combo box.