git: make "cherry picked from" suffix optional
Introduce an option "Add the 'cherry-picked from <hash>' suffix when picking commits pushed to protected branches"
It covers both cases, when users don't want to add the suffix at all,
and when they want to add a suffix to reference a "published" commit, e.g. when cherry-picking from one release branch to another.

It doesn't allow to add the suffix when picking from a feature branch,
but that seems useless, since commits may easily change in that branches, after a single rebase.

This fixes IDEA-163630.