[vcs-log] fix accidentally deleting storage files
When index files are created, if no index for the current version exists, code attempts to find and delete old index versions.
Since b2784dc825a163143ee6677282ab0cb0796cfd60 version is 10, so they have the same prefix as files for version 1.
This would not be a problem if MapIndexStorage files did not have a suffix ".storage", which was not taken into account by the code that checks storage file existence.
Solution is to add the suffix when testing and removing old storage files. See also 0c0ca5158de6e94a8c15fefeac1970dc4ec568f3

Since only MapReduceIndex files were removed, but not other maps, index is incomplete and to rebuild it index version has to be bumped.