vcs: automatically register almost all detected git/hg roots
Auto-adding a single root in simple cases proved to be convenient and caused no issues. This commit generalizes behavior to add all roots under the project dir, and the root above it (unless the whole project root is a Git root).

Can be switched off via `vcs.root.auto.add`

As it were previously, Roots 'near' the project dir (i.e. not under and not above the project dir, e.g. from some content roots with linked source) are ignored (but are displayed in Settings | Version Control).

This fixes IDEA-89259 and IDEA-192333