vcs: don't display "Git Integration Enabled" notification by default
After some discussion, it was decided, that the notification is redundant:
we found the roots, and they should just work.
This approach already exists for the "simple case" (single root equal to the project dir),
and didn't cause any issues or complaints.

Provide a registry key to change the behavior if needed, e.g. for certain IDEs.