PY-29717 In Epydoc don't add a dot at the end of each param description
Otherwise, getParamDescription(), getReturnDescription() always return
non-empty string and we can't use it do determine whether a parameter
was actually mentioned in a docstring. Besides, it's inconsistent with
other docstring formats and really unnecessary since we no longer
display types and descriptions in the same place.

This change effectively reverts PY-7006.