IDEA-192122 vcs: don't propose sync control for submodules
Submodules are normally independent from the main repository,
and sync control has little sense with them,
even if they happen to be all on the same branch at the time of the first Git Branches invocation.

Make some helper methods static, to be able to use without instantiating a DvcsMultiRootBranchConfig.