PY-30359 PY-11425 Render information about keyword arguments in sections
Similarly to normal parameters we don't show those keyword arguments
that are not explicitly present in function signature and can't be
passed in because of absence of **kwargs parameter.

Additionally, I enforced the order in which sections appears, so that
information about parameters, return value and raised exceptions
was always first and also fixed the way we show single star parameter
in signatures.