Specify updates URLs for ASWB Mac at build time
Currently, ADRT handles setting custom values for idea.updates.url and
idea.patches.url in idea.properties for ASWB on Mac builds. ADRS will
not support updating build artifacts in this manner, and so we should
set these properties at build time instead.

This CL accomplishes this by overriding createMacCustomizer() in
AndroidStudioProperties.groovy so that it returns an ASWB on
Mac-specific customizer that specifies the updates and patches URLs.

Bug: 75321718
Test: test_studio.py
Change-Id: I155021dcedeacf3a423ff6f1f833fad561fa5285