Update PluginManagerCore to use bundled Kotlin plugin
Because we now depend on the Kotlin plugin, which is still used in
binary form, when running in "internal" mode or running tests with
a single classloader, we have to also load the Kotlin code in there.

When running in different classloaders, we need to ensure that types
appearing in method signatures of platform and Kotlin plugin APIs
are loaded in the platform classloader (because we have multiple copies
of kotlin-stdlib, which leads to having incompatible definitions for
KClass etc).

Test: existing
Bug: N/A

Change-Id: Id2afc6a0bb7f08cd4c5e5cd8a41a9cf7be6de905