PY-29977 Don't access external doc URLs to validate them in getUrlFor()
This method is used both to check possible sources of external
documentation in DefaultDocumentationCollector#getDocumentation()
(executed inside a read action) and to show links
at the end of Quick Documentation in DocumentationComponent
(happens on EDT). In both cases network access may badly affect editor
responsiveness. The correct approach here is to provide the most
accurate URL right away in PythonDocumentationMap and
PythonDocumentationProvider implementations instead of falling back to
the documentation root.

As a result, PythonDocumentationLinkProvider#getExternalDocumentationURL
is no longer necessary and thus deprecated.