vcs: restore ability to Accept Yours/Theirs for submodule: IDEA-193788
MergeSession.canMerge() is intended to be used only to enable/disable the Merge button. Although this is not entirely clear from the method name, it is stated in the method's javadoc, and it is safe to behave this way, since it actually makes sense just for Git submodules.

Was broken in d3c7fa5