Have common Icons for nodes in build view
When there are multiple builds on the build view, each one of them get
its Icon from AbstractViewManager#getIcon, but the icons used are not
consistent with the ones used in ExecutionNode.

This change causes getIcon to use a new function (getEventResultIcon)
that shares the same icons as ExecutionNode#getCurrentIcon. Also, the
icons are defined using constants to cause any change on them to be used
in both functions.

(Cherry picked from commit 23b13ba)

Test: N/A
Bug: 77278972
Change-Id: Ia0bc4403956f76d16c6a7add2b3394cceb05c6a8