Fix exception due to negative values in HdrHistogram
The possible causes would be the user reducing their system clock due
to daylight savings time or a network time sync, which would be legitimate.
We deal with this case by clamping the time to 0.

Another possible cause would be an invalid argument passed to
RecursiveStopwatch.end. We deal with that case by checking for it and
throwing an IllegalArgumentException.

Bug: 110113356
Test: RecursiveStopwatchTest.testBadSystemClock
Change-Id: I0c1108ac958d7d8f682c1f488aee1ddc1bd04e2a