git: simplify the rebase line listener: used only for progress now
It fixes IDEA-194024 which was introduced by 116c1b6: when editor is cancelled, Git produces an error messages (which made the Status to be ERROR), but it is not a fatal error, and it is followed by a message about "stopped for editing". Receiving two status messages was unexpected by the GitRebaseLineListener.