git: fix handling different kinds of cancellation of interactive rebase
Cancelling the commit list editor should just break the rebase process, since nothing has actually started yet.
On the other hand, cancelling the commit message editor (e.g. during squash or reword) is different, it should put rebase into the SUSPENDED state (the same as when it stops for editing or to resolve conflicts).

Previously it was not possible to differentiate these cancellations, now a new API is introduced for this.

This fixes IDEA-194033.

It is also possible to cancel the rebase command during its execution (by pressing the cross button at the background progress), but it is handled separately in GitHandler, and throws a PCE, which is expected. This case is not touched here.