Cause warning icon to be used when there are warnings on SyncIssues
SuccessResult has a getWarnings but this method is not used by the build
view yet, causing warnings generated from SyncIssues to not show if
SuccessResult is used. As a workaround FailureResult is used, but that
causes the wrong icon to appear.

This CL checks if only warnings are present and then picks the correct icon
accordingly.

This CL will likely be reverted once getWarnings is used by the build view.

Bug: 77278972
Test: Manualy checked the warning icon is shown instead of error.
Change-Id: I91832ff8f549414a74ffc7ec97e6a57504209961