PY-30540 Associate new Python environment with new project immediately
We have enough data to associate a new Python environment with a new
project being created since we know the path to it. It is required
for Pipenv in order to be able to locate the project for creating a
virtualenv for it and a Pipfile inside it. PyCharm installs Django
during project creation so we need Pipenv to function correctly before
the project is set up.

(cherry picked from commit ee4680fa75d09d4943fdacd31290fbd4eb3116fb)