git: check new children recursively to find .git: IDEA-194449
Previously, when a new directory was created, only the directory itself was checked to be .git, but its children were not. Thus e.g. adding a submodule was not detected since .git was under new "sub" folder.

To detect such cases we need to process new directories recursively. It introduces work duplication with the VcsRootDetector which will do basically the same during the scan, but eliminating this duplication is hard in current architecture, so will be done separately.

To lessen the performance impact, process VFS events asynchronously.

(cherry picked from commit 128c9ac815ef1b46640406a3c3e963aa8782f672)