git: don't fail to update whole project is some repo can't be updated
For synchronously controlled roots it makes sense to update the project as a whole, so previous behavior (total failure if one of roots is in detached HEAD state or have no tracked branch to update from) is kept for such projects.

However, it is common for non-synced projects (including projects with submodules) to have roots checked out on detached HEAD, such are now just skipped from update.

IDEA-124453, IDEA-193934