Resolve dependency sources in all repositories.
As reported here: https://discuss.gradle.org/t/idea-plugin-isnt-downloading-sources-when-mavenlocal-is-used/2160 library sources are not downloaded for certain gradle configurations. 

Download source fails when the first maven repository configured doesn't already contain the artifact requested, a common case of this issue is use of mavenLocal which likely does not already have sources present.