PY-25488 Preserve formatting of extracted collection literals
as it's especially tedious to re-format them each time afterwards and
being surrounded with brackets they are safe to be moved to RHS of an
assignment.

Alternatively, we could leave such inlining only for binary expressions
since they can become syntactically illegal when moved out of the
context with implicit line joining (PY-2862).